rename: hooks 폴더 안에 커스텀 파일들을 다 모아넣음 - src/components/modal/useBasicModal.tsx -> src/hooks/useBasicModal.tsx - src/pages/login/useLoginPage.tsx -> src/hooks/useLoginPage.tsx - src/pages/login/useLogout.tsx -> src/hooks/useLogout.tsx - src/components/nav/useNavbar.tsx -> src/hooks/useNavbar.tsx
Github issue #121
