+++ src/pages/Calendar/AddScheduleModal.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { useSelector } from 'react-redux';
+import { RootState } from '../../store/store';
 
 interface ISchedule {
   dateId: number;
