+++ src/pages/login/LoginPage.tsx
@@ -1,26 +1,37 @@
 import * as Styled from './LoginPage.styled';
-import { useLoginPage } from './useLoginPage';
+import { useLoginPage } from '../../hooks/useLoginPage';
 import LogoImg from '../../styles/images/sajoLogo.png';
 import Btn from '../../components/button/Button';
 import { TextField } from '@mui/material';
 import { Controller } from 'react-hook-form';
 
 export default function LoginPage() {
-  const loginPage = useLoginPage(); // TODO 하이오더 컴포넌트 필요
+  const {
+    handleSubmit,
+    handleLogin,
+    control,
+    errors,
+    emailHelperText,
+    passwordHelperText,
+    checkEmailEmpty,
+    checkPasswordEmpty,
+    isLoginButtonDisabled,
+    loading,
+  } = useLoginPage();
 
   return (
     <div>
       <Styled.LoginPageContainer>
         <img src={LogoImg} alt="logo" />
         <div className="form__descriptions">{/* <h3>이메일과 비밀번호를 입력해주세요</h3> */}</div>
-        <form onSubmit={loginPage.handleSubmit(loginPage.handleLogin)}>
+        <form onSubmit={handleSubmit(handleLogin)}>
           <Controller
             name="email"
-            control={loginPage.control}
+            control={control}
             rules={{
               required: true,
               pattern: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
-              onChange: loginPage.checkEmailEmpty,
+              onChange: checkEmailEmpty,
             }}
             render={({ field }) => (
               <TextField
@@ -28,19 +39,20 @@ export default function LoginPage() {
                 className="form__login_Email"
                 {...field}
                 variant="standard"
-                helperText={loginPage.emailHelperText()}
+                helperText={emailHelperText()}
                 autoComplete="on"
-                error={loginPage.errors.email ? true : false}
+                error={errors.email ? true : false}
+                onLoad={checkEmailEmpty}
               />
             )}
           />
           <Controller
             name="password"
-            control={loginPage.control}
+            control={control}
             rules={{
               required: true,
               minLength: 8,
-              onChange: loginPage.checkPasswordEmpty,
+              onChange: checkPasswordEmpty,
             }}
             render={({ field }) => (
               <TextField
@@ -48,17 +60,18 @@ export default function LoginPage() {
                 label="비밀번호"
                 {...field}
                 variant="standard"
-                helperText={loginPage.passwordHelperText()}
+                helperText={passwordHelperText()}
                 autoComplete="on"
                 type="password"
-                error={loginPage.errors.password ? true : false}
+                error={errors.password ? true : false}
+                onLoad={checkPasswordEmpty}
               />
             )}
           />
           <Btn
-            label={`${loginPage.loading ? '로그인 중입니다..,' : '로그인'}`}
+            label={`${loading ? '로그인 중입니다..,' : '로그인'}`}
             type="submit"
-            className={loginPage.isLoginButtonDisabled}
+            className={isLoginButtonDisabled}
           ></Btn>
         </form>
       </Styled.LoginPageContainer>
