+++ src/pages/myPage/MyPage.tsx
@@ -1,6 +1,6 @@
 import Btn from '../../components/button/Button';
 import ProfileCard from '../Dashboard/ProfileCard';
-import { useLogout } from '../login/useLogout';
+import { useLogout } from '../../hooks/useLogout';
 
 export function MyPage() {
   const { handleLogout } = useLogout(); // TODO 하이오더 컴포넌트 필요
